+++ src/posts/posts.service.ts
@@ -1,12 +1,19 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Inject,
+  Injectable,
+  NotFoundException,
+  forwardRef,
+} from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PostsRepository } from './posts.repository';
 
 @Injectable()
 export class PostsService {
-  constructor(private readonly repository:PostsRepository) {
-  }
+  constructor(
+    private readonly repository: PostsRepository,
+  ) {}
   async create(body: CreatePostDto) {
     return this.repository.create(body);
   }
@@ -17,19 +24,22 @@ export class PostsService {
 
   async findOne(id: number) {
     const post = await this.repository.findOne(id);
-    if(!post) throw new NotFoundException("Post not found!")
-    return post;
+    if (!post) throw new NotFoundException('Post not found!');
+    const { medias, ...postInfo } = post;
+    return postInfo;
   }
 
   async update(id: number, body: CreatePostDto) {
     const post = await this.repository.findOne(id);
-    if(!post) throw new NotFoundException("Post not found!")
+    if (!post) throw new NotFoundException('Post not found!');
     return this.repository.update(id, body);
   }
 
   async remove(id: number) {
     const post = await this.repository.findOne(id);
-    if(!post) throw new NotFoundException("Post not found!")
+    if (!post) throw new NotFoundException('Post not found!');
+    //se existir relação com media, já foi publicado
+    if(post.medias.length!==0) throw new ForbiddenException('Already posted!')
     return this.repository.remove(id);
   }
 }
